skip xcm during migrations
