+++ primitives/xcm/src/xcm_execution_traits.rs
@@ -12,34 +12,19 @@
 // GNU General Public License for more details.
 
 // You should have received a copy of the GNU General Public License
-// along with Moonbeam.  If not, see <http://www.gnu.org/licenses/>.
 
-//! The XCM primitive trait implementations
+use sp_runtime::DispatchResult;
 
-#![cfg_attr(not(feature = "std"), no_std)]
-
-mod asset_id_conversions;
-pub use asset_id_conversions::*;
-
-mod barriers;
-pub use barriers::*;
-
-mod fee_handlers;
-pub use fee_handlers::*;
-
-mod location_conversion;
-pub use location_conversion::*;
-
-mod origin_conversion;
-pub use origin_conversion::*;
-
-mod transactor_traits;
-pub use transactor_traits::*;
-
-mod ethereum_xcm;
-pub use ethereum_xcm::*;
-
-mod xcm_execution_traits;
-pub use xcm_execution_traits::*;
-
-pub type XcmV2Weight = xcm::v2::Weight;
+/// Pause and resume execution of XCM
+pub trait PauseXcmExecution {
+	fn suspend_xcm_execution() -> DispatchResult;
+	fn resume_xcm_execution() -> DispatchResult;
+}
+impl PauseXcmExecution for () {
+	fn suspend_xcm_execution() -> DispatchResult {
+		Ok(())
+	}
+	fn resume_xcm_execution() -> DispatchResult {
+		Ok(())
+	}
+}
